+++ R/data.R
@@ -1,44 +1,26 @@
 #' Classification of instructional programs (CIP) data
 #'
-#' A dataset of codes and names for 1584 instructional programs organized on
-#' three levels: a 2-digit series, a 4-digit series, and a 6-digit series.
+#' A dataset of codes and names for 1584 instructional programs organized on three levels: a 2-digit series, a 4-digit series, and a 6-digit series.
 #'
-#' 6-digit codes are used in the midfieldr datasets (\code{midfieldstudents},
-#' \code{midfieldcourses}, \code{midfieldterms}, and  \code{midfielddegrees})
-#' to encode program information as reported by the member institutions. In
-#' addition to the standard codes and names, the midfieldr taxonomy includes
-#' three non-IPEDS codes:
+#' 6-digit codes are used in the midfieldr datasets (\code{midfieldstudents}, \code{midfieldcourses}, \code{midfieldterms}, and  \code{midfielddegrees}) to encode program information as reported by the member institutions. In addition to the standard codes and names, the midfieldr taxonomy includes three non-IPEDS codes:
 #'
 #' \describe{
 #'   \item{14XXXX}{First-Year Engineering (FYE).}
-#'   \item{14YYYY}{Engineering-Focused Curricula (pseudo-FYE). For instances
-#'   in which a student is eligible for admission to the institution but has
-#'   not been admitted to an engineering program. Some institutions without
-#'   FYE programs use such a designation in lieu of undecided or unspecified.}
-#'   \item{999999}{Undecided or Unspecified. For instances in which
-#'   institutions reported no program information or that students were
-#'   not enrolled in a program.}
+#'   \item{14YYYY}{Engineering-Focused Curricula (pseudo-FYE). For instances in which a student is eligible for admission to the institution but has not been admitted to an engineering program. Some institutions without FYE programs use such a designation in lieu of undecided or unspecified.} \item{999999}{Undecided or Unspecified. For instances in which institutions reported no program information or that students were not enrolled in a program.}
 #' }
 #'
-#' @format A data frame (tibble) with 1584 observations and 6 variables.
-#' All variables are characters. An observation is a unique program.
+#' @format A data frame (tibble) with 1584 observations and 6 variables. All variables are characters. An observation is a unique program.
 #'
 #' \describe{
-#'   \item{cip2}{An instructional program's 2-digit code,  representing "the
-#'   most general groupings of related programs" (US National Center for
-#'   Education Statistics).}
+#'   \item{cip2}{An instructional program's 2-digit code,  representing "the most general groupings of related programs" (US National Center for Education Statistics).}
 #'   \item{cip2name}{Name of a program grouping at the 2-digit level.}
-#'   \item{cip4}{An instructional program's 4-digit code, representing
-#'   "intermediate groupings of programs that have comparable content and
-#'   objectives".}
+#'   \item{cip4}{An instructional program's 4-digit code, representing "intermediate groupings of programs that have comparable content and objectives".}
 #'   \item{cip4name}{Name of a program grouping at the 4-digit level.}
-#'   \item{cip6}{An instructional program's 6-digit code, representing
-#'   "specific instructional programs".}
+#'   \item{cip6}{An instructional program's 6-digit code, representing "specific instructional programs".}
 #'   \item{cip6name}{Name of a specific program at the 6-digit level.}
 #' }
 #'
-#' @source Based on the 2010 codes curated by the US National Center for
-#' Education Statistics (\url{https://nces.ed.gov/ipeds/cipcode}).
+#' @source Based on the 2010 codes curated by the US National Center for Education Statistics (\url{https://nces.ed.gov/ipeds/cipcode}).
 #'
 #' @examples
 #' # View the CIP data
@@ -65,22 +47,18 @@
 #' @format An atomic character vector of 6-digit CIP codes.
 #'
 #' \describe{
-#' \item{\code{cip_bio_sci}}{Biological and biomedical science programs,
-#' series 26.}
+#' \item{\code{cip_bio_sci}}{Biological and biomedical science programs, series 26.}
 #' \item{\code{cip_engr}}{Engineering programs, series 14.}
-#' \item{\code{cip_math_stat}}{Mathematics and statistics programs,
-#'  series 27.}
+#' \item{\code{cip_math_stat}}{Mathematics and statistics programs, series 27.}
 #' \item{\code{cip_phys_sci}}{Physical science programs, series 40.}
 #' \item{\code{cip_other_stem}}{Other STEM programs, not individually named.}
-#' \item{\code{cip_stem}}{All Science, Technology, Engineering, and Mathematics
-#'   (STEM) programs.}
+#' \item{\code{cip_stem}}{All Science, Technology, Engineering, and Mathematics (STEM) programs.}
 #' }
 #'
 #' @source
 #' NSF STEM programs, \url{https://www.ice.gov/sites/default/files/documents/Document/2016/stem-list.pdf}
 #'
-#' CIP 2010, US National Center for Education Statistics,
-#'   \url{https://nces.ed.gov/ipeds/cipcode}.
+#' CIP 2010, US National Center for Education Statistics, \url{https://nces.ed.gov/ipeds/cipcode}.
 #'
 #' @examples
 #' head(cip_math_stat, n = 10L)
@@ -110,20 +88,14 @@ NULL
 
 #' CIP case data for four programs
 #'
-#' A subset of the [\code{cip}](cip.html) data frame for four
-#' engineering programs.
+#' A subset of the [\code{cip}](cip.html) data frame for four engineering programs.
 #'
-#' The 6-digit CIP codes and names plus a user-selected program name for Civil,
-#' Electrical, Industrial, and Mechanical Engineering. The "Select and label
-#' programs" vignette describes how these data were assembled. The data frame is
-#' used by several other vignettes as a starting point.
+#' The 6-digit CIP codes and names plus a user-selected program name for Civil, Electrical, Industrial, and Mechanical Engineering. The "Select and label programs" vignette describes how these data were assembled. The data frame is used by several other vignettes as a starting point.
 #'
-#' @format A data frame (tibble) with 12 observations and 3 variables.
-#' All variables are characters. An observation is a unique program.
+#' @format A data frame (tibble) with 12 observations and 3 variables. All variables are characters. An observation is a unique program.
 #'
 #' \describe{
-#'   \item{cip6}{An instructional program's 6-digit code, representing
-#'   "specific instructional programs".}
+#'   \item{cip6}{An instructional program's 6-digit code, representing "specific instructional programs".}
 #'   \item{cip6name}{Name of a program at the 6-digit level}
 #'   \item{program}{A program name used for grouping, summarizing, and graphing.}
 #' }
@@ -138,12 +110,9 @@ NULL
 
 #' Stickiness case data
 #'
-#' An data frame of stickiness data grouped by program and student
-#' race/ethnicity and sex.
+#' An data frame of stickiness data grouped by program and student race/ethnicity and sex.
 #'
-#' For Civil, Electrical, Industrial, and Mechanical Engineering, the number
-#' of students ever enrolled, the number of graduates, and the resulting
-#' stickiness.
+#' For Civil, Electrical, Industrial, and Mechanical Engineering, the number of students ever enrolled, the number of graduates, and the resulting stickiness.
 #'
 #' @format A data frame (tibble) with 32 observations and 6 variables.
 #'
@@ -151,12 +120,9 @@ NULL
 #'  \item{program}{Instructional programs selected for study.}
 #'  \item{race}{Student race and sex from \code{midfieldstudents} data.}
 #'  \item{sex}{Student sex from \code{midfieldstudents} data.}
-#'  \item{ever}{Number of students ever enrolled in a program. Grouped and
-#'  summarized from \code{midfieldterms} data.}
-#'  \item{grad}{Number of students graduating from the program. Grouped and
-#'  summarized from \code{midfielddegrees} data.}
-#'  \item{stick}{Program stickiness: the ratio of the number of students
-#'  graduating from a program to the number ever enrolled in the program.}
+#'  \item{ever}{Number of students ever enrolled in a program. Grouped and summarized from \code{midfieldterms} data.}
+#'  \item{grad}{Number of students graduating from the program. Grouped and summarized from \code{midfielddegrees} data.}
+#'  \item{stick}{Program stickiness: the ratio of the number of students graduating from a program to the number ever enrolled in the program.}
 #' }
 #'
 #' @source \href{https://midfieldr.github.io/midfielddata/}{midfielddata} package
@@ -171,15 +137,9 @@ NULL
 
 #' Imputed starting majors for FYE students
 #'
-#' Imputed starting majors for nearly 6000 first-year engineering (FYE) students
-#' in the midfielddata package.
+#' Imputed starting majors for nearly 6000 first-year engineering (FYE) students in the midfielddata package.
 #'
-#' Some US institutions have first year engineering (FYE) programs---typically
-#' a common first year curriculum that is a prerequisite for declaring an
-#' engineering major. For some persistence metrics, we have to determine what
-#' we call a "starting major" for FYE students---the predicted engineering
-#' program such as Civil, Electrical, or Mechanical, the student would have
-#' declared had they not been required to enroll in FYE.
+#' Some US institutions have first year engineering (FYE) programs---typically a common first year curriculum that is a prerequisite for declaring an engineering major. For some persistence metrics, we have to determine what we call a "starting major" for FYE students---the predicted engineering program such as Civil, Electrical, or Mechanical, the student would have declared had they not been required to enroll in FYE.
 #'
 #' @format A data frame (tibble) with 5992 observations and 2 variables.
 #'