+++ R/grad_filter.R
@@ -1,4 +1,4 @@
-#' @importFrom dplyr %>%  filter ungroup is.tbl
+#' @importFrom dplyr filter ungroup is.tbl
 #' @importFrom stringr str_c str_detect
 #' @importFrom tidyr drop_na
 #' @importFrom wrapr stop_if_dot_args let
@@ -6,54 +6,38 @@ NULL
 
 #' Filter degree data for graduating students
 #'
-#' Filter academic degree data to find all students graduating from specified
-#' programs.
+#' Filter academic degree data to find all students graduating from specified programs.
 #'
-#' To use this function, the \code{midfielddata} package must be installed to
-#' provide \code{midfielddegrees}, the default reference data set.
+#' The \code{midfielddata} package must be installed to provide the \code{midfielddegrees} data frame as the default value for the \code{data} argument. Optionally, a data frame having the same structure as \code{midfielddegrees} may be used as the value for the \code{data} argument.
 #'
-#' The \code{series} argument is an atomic character vector of 6-digit CIP
-#' codes used to filter the reference data set and extract student IDs and terms.
+#' The \code{filter_by} argument is an atomic character vector of 6-digit CIP codes used to filter \code{data} and extract student IDs and terms.
 #'
-#' The function returns a subset of \code{reference} with every unique
-#' combination of student ID and CIP of the program(s) in which they earned
-#' their first degree(s). Only the student ID and program CIP are returned;
-#' other variables in \code{reference} are quietly dropped.
+#' The function returns a subset of \code{data} with every unique combination of student ID and CIP of the program(s) in which they earned their first degree(s). Only the student ID and program CIP are returned; other variables in \code{data} are quietly dropped.
 #'
-#' Optional arguments. An alternate reference data set can be assigned
-#' via the \code{reference} argument. The alternate must have variables for
-#' student ID, 6-digit CIP, and terms in which degrees were earned with the
-#' same structure as in \code{midfielddegrees}. Use the  \code{id}, \code{cip6},
-#' and \code{term} arguments to reassign the variables names, if necessary, to
-#' match the variable names in the alternate.
+#' Optional arguments. Student ID variables in the midfielddata data sets are named "id". If your data frames use a different name, you can either 1) rename your variables to "id", or 2) use the optional \code{id} argument to pass the alternate variable name to the function. The same is true for the \code{cip6} and \code{term} variables.
 #'
-#' @param series atomic character vector specifying the 6-digit CIP codes of
-#' the programs to filter by
+#' @param data Data frame of student IDs, academic terms, and CIP codes, default \code{midfielddegrees}.
 #'
-#' @param ... not used for values, forces later arguments to bind by name
+#' @param filter_by Atomic character vector of 6-digit CIP codes specifying the programs to filter by.
 #'
-#' @param reference a reference data frame from which graduating student IDs,
-#' CIP codes, and terms are obtained, default \code{midfielddegrees}
+#' @param ... Not used for values, forces later arguments to bind by name
 #'
-#' @param id character column name of the ID variable in \code{reference}
+#' @param id Optional argument, the quoted column name of the student ID variable in \code{data}. Default is "id".
 #'
-#' @param cip6 character column name of the CIP code variable in \code{reference}
+#' @param cip6 Optional argument, the quoted column name of the 6-digit CIP code variable in \code{data}. Default is "cip6".
 #'
-#' @param term character column name of term variable in \code{reference}
+#' @param term Optional argument, the quoted column name of the term variable in \code{data}. Default is "term".
 #'
 #' @return Data frame with character variables for student ID and program CIP
 #'
-#' @seealso \code{\link[midfieldr]{cip_filter}} for obtaining 6-digit CIP codes
+#' @seealso \code{\link[midfieldr]{cip6_select}} for selecting 6-digit CIP codes and naming programs.
 #'
 #' @examples
-#' grad <- grad_filter(series = "540104")
-#' grad
+#' library("midfielddata")
+#' (grad_filter(filter_by = "540104"))
+#'
 #' @export
-grad_filter <- function(series, ...,
-                        reference = NULL,
-                        id = "id",
-                        cip6 = "cip6",
-                        term = "term_degree") {
+grad_filter <- function(data = NULL, filter_by = NULL, ..., id = "id", cip6 = "cip6", term = "term_degree") {
   if (!.pkgglobalenv$has_data) {
     stop(paste(
       "To use this function, you must have",
@@ -64,27 +48,25 @@ grad_filter <- function(series, ...,
   # force optional arguments to be usable only by name
   wrapr::stop_if_dot_args(substitute(list(...)), "grad_filter")
 
-  if (is.null(series)) {
-    stop("midfieldr::grad_filter, series missing or incorrectly specified")
+  # argument checks
+  if (is.null(data)) {
+    data <- midfielddata::midfielddegrees
+  }
+  if (!(is.data.frame(data) || dplyr::is.tbl(data))) {
+    stop("midfieldr::grad_filter, data must be a data frame or tbl")
+  }
+  if (is.null(filter_by)) {
+    stop("midfieldr::grad_filter, filter_by cannot be NULL")
+  }
+  if (isFALSE(is.atomic(filter_by))) {
+    stop("midfieldr::grad_filter, filter_by must be an atomic variable")
   }
 
-  if (isFALSE(is.atomic(series))) {
-    stop("midfieldr::grad_filter, series must be an atomic variable")
-  }
-
-  # assign the default terms dataset
-  if (is.null(reference)) {
-    reference <- midfielddata::midfielddegrees
-  }
-
-  if (!(is.data.frame(reference) || dplyr::is.tbl(reference))) {
-    stop("midfieldr::grad_filter, reference must be a data frame or tbl")
-  }
-
-  collapse_series <- stringr::str_c(series, collapse = "|")
+  # search terms in a single string
+  collapse_series <- stringr::str_c(filter_by, collapse = "|")
 
   # addresses R CMD check warning "no visible binding"
-  ID <- NULL
+  ID   <- NULL
   CIP6 <- NULL
   TERM <- NULL
 
@@ -94,7 +76,7 @@ grad_filter <- function(series, ...,
     alias = mapping,
     expr = {
       # filter for data for specified programs
-      students <- dplyr::select(reference, ID, CIP6, TERM)
+      students <- dplyr::select(data, ID, CIP6, TERM)
       students <- dplyr::filter(
         students,
         stringr::str_detect(CIP6, collapse_series)