+++ R/multiway_order.R
@@ -9,26 +9,15 @@ NULL
 
 #' Prepare multiway data for graphing
 #'
-#' Transforms a data frame such that two categorical variables become factors
-#' with their levels ordered by medians of the quantitative variable.
+#' Transforms a data frame such that two categorical variables become factors with their levels ordered by medians of the quantitative variable.
 #'
-#' In multiway data, there is a single quantitative value (or response) for
-#' every combination of levels of two categorical variables.
-#' \code{multiway_order()} converts the categorical variables to factors and
-#' orders the factors by increasing median values of the levels of the
-#' categorical variables.
+#' In multiway data, there is a single quantitative value (or response) for every combination of levels of two categorical variables. \code{multiway_order()} converts the categorical variables to factors and orders the factors by increasing median values of the levels of the categorical variables.
 #'
-#' @param mw_data A data frame with one numerical variable and two categorical
-#' variables.
+#' @param mw_data Data frame with one numerical variable and two categorical variables.
 #'
-#' @param return_medians A logical value indicating whether or not the medians
-#' computed to order the factors should be included in the data frame
-#' returned. The default is FALSE.
+#' @param return_medians Logical value indicating whether or not the medians computed to order the factors should be included in the data frame returned. The default is FALSE.
 #'
-#' @return A data frame with its two categorical variables converted
-#' to factors and ordered by increasing medians of the numerical variable. The
-#' medians of the quantitative variable for each category are added to the
-#' data frame if argument \code{medians} is TRUE.
+#' @return Data frame with its two categorical variables converted to factors and ordered by increasing medians of the numerical variable. The medians of the quantitative variable for each category are added to the data frame if argument \code{medians} is TRUE.
 #'
 #' @examples
 #' # construct a data frame with one numeric and two character variables
@@ -48,6 +37,7 @@ NULL
 #' df3 <- multiway_order(df1, return_medians = TRUE)
 #' df3 <- dplyr::arrange(df3, cat1, cat2)
 #' df3
+#'
 #' @export
 multiway_order <- function(mw_data, return_medians = FALSE) {
   if (!(is.data.frame(mw_data) || dplyr::is.tbl(mw_data))) {