+++ R/race_sex_join.R
@@ -1,39 +1,38 @@
-#' @importFrom dplyr select left_join anti_join %>%
+#' @importFrom dplyr select left_join anti_join
 #' @importFrom wrapr stop_if_dot_args let
 NULL
 
 #' Join student demographics
 #'
-#' Add student race and sex to a data frame.
+#' Add student race/ethnicity and sex variables to a data frame.
 #'
-#' To use this function, the \code{midfielddata} package must be installed to provide \code{midfieldstudents}, the default reference data set.
+#' The \code{data} argument is a data frame to which student race/ethnicity and sex data are to be joined and returned.
 #'
-#' The student IDs in \code{data} are assumed to be a subset of the IDs in  \code{reference} with the same variable name. This name can be changed to match a different reference data set using the optional \code{id} argument.
+#' Student race/ethnicity and sex information are obtained from the \code{demographics} data frame with a default value of  \code{midfieldstudents} from the midfielddata package. Optionally, any data frame having the same structure as \code{midfieldstudents} can be used.
 #'
-#' Race and sex data are obtained from the reference data set. These variable names can also be changed to match a different reference data set using the optional \code{race} and \code{sex} arguments.  These variables are joined to \code{data} using ID as the join-by variable.
+#' The student ID variables in \code{data} and \code{demographics} must have the same column name. The values of the ID variable in \code{data} are assumed to be a subset of the values in \code{demographics}.
 #'
-#' If sex or race variables are already present in the input data frame, they are returned unaltered.
+#' If  \code{data} already contains \code{race} and \code{sex} variables, a join does not take place and \code{data} is returned unaltered.
 #'
-#' @param data data frame to which race and sex are to be joined by student IDs
+#' Optional arguments. Student ID variables in the midfielddata data sets are named "id". If your data frames use a different name, you can either 1) rename your variables to "id", or 2) use the optional \code{id} argument to pass the alternate variable name to the function. The same is true for the \code{race} and \code{sex} variables.
 #'
-#' @param ... not used for values, forces later optional arguments to bind by name
+#' @param data Data frame to which race and sex are to be joined by student IDs
 #'
-#' @param reference a reference data frame from which student race and sex are obtained, default \code{midfieldstudents}
+#' @param demographics A data frame from which student race and sex are obtained, default \code{midfieldstudents}
 #'
-#' @param id character column name of the ID variable in both \code{data} and \code{reference}
 #'
-#' @param race character column name of the race variable in \code{reference}
+#' @param ... Not used for values, forces later optional arguments to bind by name.
 #'
-#' @param sex character column name of the sex variable in \code{reference}
+#' @param id Optional character column name of the ID variable in both \code{data} and \code{demographics}.
 #'
-#' @return The input data frame is returned with two new columns for race and sex
+#' @param race Optional character column name of the race variable in \code{demographics}.
+#'
+#' @param sex Optional character column name of the sex variable in \code{demographics}.
+#'
+#' @return The input data frame is returned with two new columns for race and sex.
 #'
 #' @export
-race_sex_join <- function(data, ...,
-                          reference = NULL,
-                          id = "id",
-                          race = "race",
-                          sex = "sex") {
+race_sex_join <- function(data = NULL, demographics = NULL, ..., id = "id", race = "race",  sex = "sex") {
 
   # force optional arguments to be usable only by name
   wrapr::stop_if_dot_args(substitute(list(...)), "race_sex_join")
@@ -45,22 +44,21 @@ race_sex_join <- function(data, ...,
     ))
   }
 
+  # argument checks
+  if (is.null(demographics)) {
+    demographics <- midfielddata::midfieldstudents
+  }
   if (!(is.data.frame(data) || dplyr::is.tbl(data))) {
-    stop("midfieldr::race_sex_join() argument must be a data.frame or tbl")
+    stop("midfieldr::race_sex_join() data argument must be a data.frame or tbl")
   }
   if (is.null(data)) {
     stop("midfieldr::race_sex_join, data cannot be NULL")
   }
 
-  # assign the default reference data set
-  if (is.null(reference)) {
-    reference <- midfielddata::midfieldstudents
-  }
-
   # addresses R CMD check warning "no visible binding"
-  ID <- NULL
+  ID   <- NULL
   RACE <- NULL
-  SEX <- NULL
+  SEX  <- NULL
 
   # use wrapr::let() to allow alternate column names
 
@@ -71,12 +69,12 @@ race_sex_join <- function(data, ...,
         stop("midfieldr::race_sex_join, id missing from data")
       }
 
-      if (isFALSE(ID %in% names(reference) ||
-        RACE %in% names(reference) ||
-        SEX %in% names(reference))) {
+      if (isFALSE(ID %in% names(demographics) ||
+        RACE %in% names(demographics) ||
+        SEX %in% names(demographics))) {
         stop(paste(
           "midfieldr::race_sex_join, id, race, or sex",
-          "missing from reference data"
+          "missing from demographics"
         ))
       }
 
@@ -84,7 +82,7 @@ race_sex_join <- function(data, ...,
       if (isTRUE(SEX %in% names(data)) && isTRUE(RACE %in% names(data))) {
         return(data)
       } else {
-        all_id_race_sex <- dplyr::select(reference, ID, RACE, SEX)
+        all_id_race_sex <- dplyr::select(demographics, ID, RACE, SEX)
       }
       if (isTRUE(race %in% names(data))) {
         all_id_race_sex <- dplyr::select(all_id_race_sex, ID, SEX)
@@ -97,7 +95,7 @@ race_sex_join <- function(data, ...,
       # are not matching values in y, keeping just columns from x.
       mismatch <- dplyr::anti_join(data, all_id_race_sex, by = ID)
       if (isFALSE(identical(nrow(mismatch), 0L))) {
-        stop("midfieldr::race_sex_join, id mismatch between data and reference")
+        stop("midfieldr::race_sex_join, id mismatch between data and demographics")
       }
 
       # join race and or sex by id