+++ R/rcb.R
@@ -6,15 +6,13 @@ NULL
 #'
 #' Accesses the \code{midfieldr::rcb_colors} dataset and returns a hex color code.
 #'
-#' @param pattern A character variable of a custom color name of the form
-#'  "level_hue",  with 4 levels (dark, mid, light, pale) and 5 hues
-#'   (Br, BG, PR, Gn, Gray or Grey). For example, "dark_Br", "mid_BG",
-#'   "light_Gn", or "pale_Gray."
+#' @param pattern A character variable of a custom color name of the form "level_hue",  with 4 levels (dark, mid, light, pale) and 5 hues (Br, BG, PR, Gn, Gray or Grey). For example, "dark_Br", "mid_BG", "light_Gn", or "pale_Gray."
 #'
 #' @return A hexadecimal color code in character form "#nnnnnn"
 #'
 #' @examples
 #' rcb("dark_Br")
+#'
 #' @export
 rcb <- function(pattern) {