+++ R/utils.R
@@ -1,7 +1,6 @@
 
 
 
-
 #' Pipe operator
 #'
 #' @name %>%
@@ -14,14 +13,15 @@ NULL
 
 
 
-# adapted from
-# G. Brooke Anderson and Dirk Eddelbuettel
-# The R Journal (2017) 9:1, pages 486-497
-# https://journal.r-project.org/archive/2017/RJ-2017-026/index.html
+
 
 .pkgglobalenv <- new.env(parent = emptyenv())
 
 .onAttach <- function(libname, pkgname) {
+
+  # G. Brooke Anderson and Dirk Eddelbuettel
+  # The R Journal (2017) 9:1, pages 486-497
+  # https://journal.r-project.org/archive/2017/RJ-2017-026/index.html
   has_data_package <- requireNamespace("midfielddata")
   if (!has_data_package) {
     packageStartupMessage(paste(
@@ -32,6 +32,12 @@ NULL
   }
   assign("has_data", has_data_package, envir = .pkgglobalenv)
 }
+
+
+
+
+
+