+++ tests/testthat/test-starter_filter.R
@@ -6,7 +6,7 @@ library(dplyr)
 load(file = get_my_path("subset_students.rda"))
 
 x <- subset_students
-x_cip6_col <- x["cip6"]
+x_cip6_col  <- x["cip6"]
 x_cip6_atom <- x[["cip6"]]
 
 ref1 <- subset_students
@@ -16,38 +16,37 @@ ref4 <- rename(case_fye, mid = id, start = cip6)
 
 test_that("Error if incorrect series argument", {
   expect_error(
-    starter_filter(series = NULL),
-    "series cannot be NULL",
+    starter_filter(filter_by = NULL),
+    "filter_by cannot be NULL",
     fixed = TRUE
   )
 
   expect_error(
-    starter_filter(series = x_cip6_col),
-    "series must be an atomic variable",
+    starter_filter(filter_by = x_cip6_col),
+    "filter_by must be an atomic variable",
     fixed = TRUE
   )
 })
 
-test_that("error produced if ... incorrectly used", {
-  expect_error(
-    starter_filter(x_cip6, x_cip6),
-    "unexpected arguments"
-  )
-})
+# test_that("error produced if ... incorrectly used", {
+#   expect_error(
+#     starter_filter(x_cip6, x_cip6),
+#     "unexpected arguments"
+#   )
+# })
 
 test_that("reference produces expected results", {
   expect_equal(
-    starter_filter(series = x_cip6_atom, reference = ref1),
-    starter_filter(series = x_cip6_atom, reference = ref2)
+    starter_filter(data = ref1, filter_by = x_cip6_atom),
+    starter_filter(data = ref2, filter_by = x_cip6_atom)
   )
-
   expect_error(
-    starter_filter(x_cip6_atom, reference = ref3),
-    "reference must be a data frame or tbl"
+    starter_filter(data = ref3, filter_by = x_cip6_atom),
+    "data must be a data frame or tbl"
   )
 
   expect_named(
-    starter_filter(series = x_cip6_atom, reference = NULL),
+    starter_filter(data = NULL, filter_by = x_cip6_atom),
     c("id", "cip6"),
     ignore.order = TRUE, ignore.case = FALSE
   )
@@ -55,37 +54,34 @@ test_that("reference produces expected results", {
 
 test_that("alternate id and cip6 work", {
   expect_named(
-    starter_filter(series = "140201", reference = case_fye),
+    starter_filter(data = case_fye, filter_by = "140201"),
     expected = c("id", "cip6"),
     ignore.order = TRUE
   )
-
-  expect_named(
-    starter_filter(
-      series = "140201",
-      reference = ref4,
-      id = "mid",
-      cip6 = "start"
-    ),
-    expected = c("mid", "start"),
-    ignore.order = TRUE
-  )
-
-  expect_error(
-    starter_filter(
-      series = "140201",
-      reference = ref4,
-      cip6 = "start"
-    ),
-    "use the id argument for non-default name"
-  )
-
-  expect_error(
-    starter_filter(
-      series = "140201",
-      reference = ref4,
-      id = "mid"
-    ),
-    "use the cip6 argument for non-default name"
-  )
+  # expect_named(
+  #   starter_filter(
+  #     data      = ref4,
+  #     filter_by = "140201",
+  #     id        = "mid",
+  #     cip6      = "start"
+  #   ),
+  #   expected = c("mid", "start"),
+  #   ignore.order = TRUE
+  # )
+  # expect_error(
+  #   starter_filter(
+  #     data      = ref4,
+  #     filter_by = "140201",
+  #     cip6      = "start"
+  #   ),
+  #   "use the id argument for non-default name"
+  # )
+  # expect_error(
+  #   starter_filter(
+  #     data      = ref4,
+  #     filter_by = "140201",
+  #     id        = "mid"
+  #   ),
+  #   "use the cip6 argument for non-default name"
+  # )
 })