+++ NEWS.md
@@ -8,7 +8,9 @@ Changes incompatible with earlier versions.
 
 - For functions that operate on a data frame and return a data frame, the data frame is now the first argument---conforming to conventional coding practice and facilitating the use of pipe operators (`%>%`). For example, the new `cip_filter(data = cip, series)` replaces the old `cip_filter(series, reference = cip)`.  
 - `cip6_select()` replaces `cip_label()`. Both functions add a custom program label for grouping, summarizing, and graphing, but the new function removes columns of 2-digit and 4-digit codes and names. 
-- `ever_filter(data, filter_by)` revised names for arguments and reordered  
+- `ever_filter(data, filter_by)` argument names edited and reordered 
+- `grad_filter(data, filter_by)` argument names edited and reordered  
+- `race_sex_join(data, demographics)` argument names edited and reordered  
 
 Additional improvements 
 
@@ -17,6 +19,7 @@ Additional improvements
 - `cip6_select(program = NULL)` assigns the `cip4name` values to the `program` variable. The old `cip_label()` NULL result was the 6-digit names. 
 - `cip6_select(program = "named_series")` is a new option for the program argument that assigns the appropriate named series to the `program` values. 
 - Content links and session information added to each vignette   
+- Added internal data bits, e.g., dimensions and memory requirements of midfielddata data frames, for use in vignettes and README